- Move most input handling to the MCGLView - Remove context holding from the Gamepad object - Simplified main_with_custom_ctrl layout hierarchy
