+++ app_pojavlauncher/src/main/res/layout/main_with_customctrl.xml
@@ -33,13 +33,6 @@
 				android:visibility="gone"/>
 
 
-			<ImageView
-				android:id="@+id/console_pointer"
-				android:visibility="invisible"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				app:srcCompat="@drawable/pointer" />
-
 			<net.kdt.pojavlaunch.customcontrols.TouchCharInput
 				android:id="@+id/editTextTextPersonName2"
 				android:layout_width="1dp"
